requirements.txt
accelerate==0.27.2
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
av==12.0.0
bitsandbytes==0.41.0
braceexpand==0.1.7
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.0.1
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
datasets==2.16.1
debugpy==1.8.1
decorator==5.1.1
decord==0.6.0
deepspeed==0.14.4
dill==0.3.7
distro==1.9.0
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.13.3
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2023.10.0
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
gradio==3.35.2
gradio_client==0.2.9
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.22.2
hydra-core==1.3.2
idna==3.6
imageio==2.35.1
importlib_metadata==7.1.0
iopath==0.1.10
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
linkify-it-py==2.0.3
lit==18.1.2
markdown-it-py==2.2.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
mmengine==0.10.4
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.3
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.8.0
opencv-python==4.8.0.74
openpyxl==3.1.2
orjson==3.10.0
packaging==24.0
pandas==2.2.1
parameterized==0.9.0
parso==0.8.4
peft==0.12.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
portalocker==2.8.2
prompt_toolkit==3.0.46
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-spy==0.3.14
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycocoevalcap==1.2
pydantic==1.10.14
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytorchvideo==0.1.5
pytz==2024.1
pywsd==1.2.4
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.12.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.44.0
setproctitle==1.3.3
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.8.30
timm==0.9.12
tokenizers==0.15.2
tomli==2.0.1
toolz==0.12.1
torch==2.0.1
torchvision==0.15.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.36.2
triton==2.0.0
typing_extensions==4.10.0
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.1
uvicorn==0.29.0
wandb==0.16.5
wavedrom==2.0.3.post3
wcwidth==0.2.13
webdataset==0.2.86
websockets==12.0
wn==0.9.5
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.18.2
s2wrapper @ git+https://github.com/bfshi/scaling_on_scales@f08aec91337ae1ed6d7cc7a55441a96d51c14dd1